requirements.txt
# Core (no dependencies needed - pure Python)

# Engine
asyncio-mqtt>=2.0

# Adapters
websockets>=12.0
python-binance>=1.0.19
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
redis[hiredis]>=5.0
python-telegram-bot>=21.0
firebase-admin>=6.0
httpx>=0.27

# API
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9

# AI Agent
openai>=1.12

# Dev & Test
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
ruff>=0.2
mypy>=1.8
